# api/openapi.yaml
components:
  schemas:
...
    ConfirmPasswordResetRequest:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - https://example.com/ConfirmPasswordResetRequest.json
          format: uri
          readOnly: true
          type: string
        code:
          description: The code delivered after requesting a reset
          maxLength: 6
          minLength: 6
          type: string
        email:
          description: Email address
          format: email
          maxLength: 255
          type: string
        password:
          description: New password
          maxLength: 72
          minLength: 12
          type: string
        password_confirm:
          description: Must match password
          maxLength: 72
          minLength: 12
          type: string
      required:
        - email
        - code
        - password
        - password_confirm
      type: object
...
    CreateSessionRequest:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - https://example.com/CreateSessionRequest.json
          format: uri
          readOnly: true
          type: string
        email:
          description: Email address
          format: email
          maxLength: 255
          type: string
        password:
          description: Plain-text password
          maxLength: 72
          minLength: 1
          type: string
      required:
        - email
        - password
      type: object
    CreateUserRequest:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - https://example.com/CreateUserRequest.json
          format: uri
          readOnly: true
          type: string
        email:
          description: Email address
          format: email
          maxLength: 255
          type: string
        name:
          description: Display name
          maxLength: 100
          minLength: 1
          type: string
        password:
          description: Plain-text password, hashed before storage
          maxLength: 72
          minLength: 12
          type: string
        password_confirm:
          description: Must match password
          maxLength: 72
          minLength: 12
          type: string
      required:
        - name
        - email
        - password
        - password_confirm
      type: object
...
    Document:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - https://example.com/Document.json
          format: uri
          readOnly: true
          type: string
        code:
          description: Stable machine-readable identifier for this refusal
          type: string
        detail:
          description: Human-readable explanation, in the negotiated language. Never parse this — use code.
          type: string
        errors:
          description: Field-level problems, when the request failed validation
          items:
            $ref: "#/components/schemas/ErrorDetail"
          type:
            - array
            - "null"
        instance:
          description: A URI identifying this occurrence
          format: uri
          type: string
        required_permission:
          description: Permission key the caller was missing
          type: string
        status:
          description: HTTP status code
          format: int64
          type: integer
        title:
          description: Short, human-readable summary of the problem type
          type: string
        type:
          description: A URI identifying the problem type
          format: uri
          type: string
      type: object
    ErrorDetail:
      additionalProperties: false
      properties:
        location:
          description: Where the error occurred, e.g. 'body.items[3].tags' or 'path.thing-id'
          type: string
        message:
          description: Error message text
          type: string
        value:
          description: The value at the given location
      type: object
...
    RequestPasswordResetRequest:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - https://example.com/RequestPasswordResetRequest.json
          format: uri
          readOnly: true
          type: string
        email:
          description: Email address
          format: email
          maxLength: 255
          type: string
      required:
        - email
      type: object
...
    SessionResponse:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - https://example.com/SessionResponse.json
          format: uri
          readOnly: true
          type: string
        device_token:
          description: Opaque secret identifying this device on later sign-ins
          type: string
        expires_at:
          description: When the token stops being accepted
          format: date-time
          type: string
        token:
          description: Bearer token for subsequent requests
          type: string
        two_factor_required:
          description: True when a code was emailed and no token was issued
          type: boolean
        user:
          $ref: "#/components/schemas/UserResponse"
      required:
        - two_factor_required
      type: object
...
    UserResponse:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - https://example.com/UserResponse.json
          format: uri
          readOnly: true
          type: string
        created_at:
          description: When the account was registered
          format: date-time
          type: string
        email:
          description: Email address, normalised to lower case
          format: email
          type: string
        id:
          description: Unique identifier
          format: uuid
          type: string
        locale:
          description: Preferred language, remembered from registration. Outranks Accept-Language.
          type: string
        name:
          description: Display name
          type: string
        two_factor_enabled:
          description: Whether sign-in from an untrusted device needs an emailed code
          type: boolean
      required:
        - id
        - name
        - email
        - two_factor_enabled
        - created_at
      type: object
    VerifySessionRequest:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - https://example.com/VerifySessionRequest.json
          format: uri
          readOnly: true
          type: string
        code:
          description: The code delivered after signing in
          maxLength: 6
          minLength: 6
          type: string
        email:
          description: Email address
          format: email
          maxLength: 255
          type: string
      required:
        - email
        - code
      type: object
  securitySchemes:
    bearer:
      bearerFormat: JWT
      scheme: bearer
      type: http
info:
  description: Tracks users, their known devices, and login history.
  license:
    identifier: MIT
    name: MIT
  title: Example
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /v1/password-resets:
    post:
      description: Sends a one-time code to the address if it is registered. The response is always 204 — including when delivery or storage fails — so the call cannot be used to discover accounts. In development without SMTP the code is written to the process log.
      operationId: request-password-reset
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestPasswordResetRequest"
        required: true
      responses:
        "204":
          description: No Content
        "422":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Document"
          description: Unprocessable Entity
        "429":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Document"
          description: Too Many Requests
        "500":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Document"
          description: Internal Server Error
      summary: Request a password reset
      tags:
        - auth
  /v1/password-resets/confirm:
    post:
      description: Spends the code and sets a new password. password and password_confirm must match. Unknown addresses, expired codes and wrong codes share one error.
      operationId: confirm-password-reset
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmPasswordResetRequest"
        required: true
      responses:
        "204":
          description: No Content
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Document"
          description: Unauthorized
        "422":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Document"
          description: Unprocessable Entity
        "429":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Document"
          description: Too Many Requests
        "500":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Document"
          description: Internal Server Error
      summary: Confirm a password reset
      tags:
        - auth
...
  /v1/sessions:
    post:
      description: Exchanges an email and password for a Bearer token. Wrong passwords and unknown addresses share one error. When the account has two-factor on and the X-Device-Token is missing or names an untrusted device, the answer is 202 with two_factor_required and a code goes out by email; finish at POST /v1/sessions/verify.
      operationId: create-session
      parameters:
        - description: Device token from a previous sign-in, if any
          in: header
          name: X-Device-Token
          schema:
            description: Device token from a previous sign-in, if any
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSessionRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
          description: Created
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Document"
          description: Unauthorized
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Document"
          description: Forbidden
        "422":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Document"
          description: Unprocessable Entity
        "429":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Document"
          description: Too Many Requests
        "500":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Document"
          description: Internal Server Error
      summary: Sign in
      tags:
        - auth
  /v1/sessions/verify:
    post:
      description: Spends the emailed code and issues the token. The X-Device-Token must be the one returned by the sign-in that raised the challenge — a code authorises one device, not one mailbox. Succeeding also trusts that device, so later sign-ins from it skip the code. Every way of getting it wrong shares one error.
      operationId: verify-session
      parameters:
        - description: Device token the code was issued for
          in: header
          name: X-Device-Token
          required: true
          schema:
            description: Device token the code was issued for
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifySessionRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
          description: Created
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Document"
          description: Unauthorized
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Document"
          description: Forbidden
        "422":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Document"
          description: Unprocessable Entity
        "429":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Document"
          description: Too Many Requests
        "500":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Document"
          description: Internal Server Error
      summary: Finish a two-factor sign-in
      tags:
        - auth
  /v1/users:
    post:
      description: Creates an account. password and password_confirm must match. A duplicate email is reported as success so the response cannot be used to discover registered addresses. Sign in with POST /v1/sessions to obtain a token.
      operationId: create-user
      parameters:
        - description: Language to remember for this account. Absent means no preference, and later requests keep negotiating per request.
          in: header
          name: Accept-Language
          schema:
            description: Language to remember for this account. Absent means no preference, and later requests keep negotiating per request.
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
        required: true
      responses:
        "204":
          description: No Content
        "422":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Document"
          description: Unprocessable Entity
        "429":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Document"
          description: Too Many Requests
        "500":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Document"
          description: Internal Server Error
      summary: Register a user
      tags:
        - auth
...